pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.5.2",
    "aiohttp>=3.9.5",
    "beautifulsoup4>=4.12.3",
    "spacy>=3.7.2",
    "flask>=2.3.3"
]
